adding some tags
Former-commit-id: 00bd5f21a4a90aefea19231941e4e8124278a9c4

+++ src/Game/Data.hs
@@ -44,5 +44,5 @@ data Zone = Zone { grid :: [Int]
 
 data Unit = Unit { unittype :: Int
                  , zone     :: (Int, Int)
-                 , position :: (Int, Int)
+                 , position :: (Float, Float)
                  } deriving (Show, Eq)
